+++ changelog.md
@@ -6,6 +6,7 @@
 - Adset card added to TTS json download
 - Crafted improvements tile added to TTS json
 - Double sided Pieces on Board Back
+- Tokens and Buildings added to TTS json
 
 ### Improvements
 - Delete button for Forged Factions
@@ -13,12 +14,13 @@
 - Replaced crafted items svg with dynamicly created box
 - Dynamic min width for attribute bars
 - Soft warning if name in Forge conflicts with existing faction
+- Pieces snap to matching tracks
 
 ### Bug Fixes
 - Piece name field is now optional
 - Fixed (none) text to be black and standard font
 - Adset number and text alignment adjusted
-- Lock faction board after 1 sec in TTS to work around floating boards
+- Lock faction board after 1.5 sec in TTS to work around floating boards
 
 ## [1.12.1] - 2026-5-05 Forge-Tuning